feat(accounts): add edit account feature with dropdown actions and zustand
- Added Actions component for account edit dropdown
- Introduced zustand state management for editing account (useOpenAccount)
- Created edit-account-sheet component for account edit form
- Integrated edit account feature in accounts/page.tsx
- Added useGetAccount hook to fetch account data by ID
- Updated data-table and sheet-provider components to support the new feature
- Added UI dropdown-menu component

+++ components/data-table.tsx
@@ -12,6 +12,8 @@ import {
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table"
+
+import { useConfirm } from "@/hooks/use-confirm"
 import { Button } from "./ui/button"
 import { Input } from "@/components/ui/input"
 
@@ -40,6 +42,11 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmDialog, confirm] = useConfirm(
+    "Are you sure?",
+    "You are about to perform a bulk delete."
+  );
+
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -66,6 +73,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
         <div className="flex items-center py-4">
           <Input
             placeholder={`Filter ${filterKey}...`}
@@ -81,9 +89,13 @@ export function DataTable<TData, TValue>({
               size='sm'
               variant='outline'
               className="ml-auto font-normal text-xs"
-              onClick={() => {
-                onDelete(table.getFilteredSelectedRowModel().rows)
-                table.resetRowSelection();
+              onClick={ async () => {
+                const ok = await confirm();
+
+                if (ok) {
+                  onDelete(table.getFilteredSelectedRowModel().rows)
+                  table.resetRowSelection();
+                }
               }}
             >
               <Trash  className="size-4 mr-2"/>
